// IOITQln/web-app/src/app/core/models/check-role-model.ts
export class CheckRoleModel {
    ViewOrActionSpecial: boolean = false;
    Create: boolean = false;
    Update: boolean = false;
    Delete: boolean = false;
    Import: boolean = false;
    Export: boolean = false;
    Print: boolean = false;
    Other: boolean = false;
    Menu: boolean = false;
}